scanner/requirements.txt
google-auth==2.32.0
google-cloud-spanner==3.47.0
google-cloud-resource-manager==1.12.4
google-api-python-client==2.137.0
plotly==5.22.0
networkx==3.3
numpy==2.0.0
json2html==1.3.0
oauth2client==4.1.3
